% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4app-package.R
\docType{package}
\name{c4app-package}
\alias{c4app}
\alias{c4app-package}
\title{c4app: c4app}
\description{
PKG_DESC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shaenzi/R_course_04_project}
  \item Report bugs at \url{https://github.com/shaenzi/R_course_04_project/issues}
}

}
\author{
\strong{Maintainer}: Sara Hänzi \email{AUTHOR@MAIL.COM}

}
\keyword{internal}
