% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zurich_kreise}
\alias{zurich_kreise}
\title{Urban districts ("Stadtkreise" ) of Zurich}
\format{
## `zurich_kreise`
\describe{
  \item{bezeichnun}{Name of the district, "Kreis 1" to "Kreis 12"}
  \item{geometry}{polygon to be plotted, borders of the urban district}
  }
}
\source{
<https://www.stadt-zuerich.ch/geodaten/download/Stadtkreise?format=10007>
}
\usage{
zurich_kreise
}
\description{
Geometries/polygons of the urban districts of Zurich, downloaded from the open government data portal
}
\keyword{datasets}
